Some modif using all caplog
